Fix bug with starting up
